Split render system types into new header
This splits several enums and a typedef  into a new header file,
RenderSystemTypes.h. This avoids a dependence on the base rendering system
class if external code only uses the type.
