+++ xbmc/rendering/RenderSystem.h
@@ -23,10 +23,10 @@
 
 #pragma once
 
+#include "RenderSystemTypes.h"
 #include "guilib/Geometry.h"
 #include "guilib/TransformMatrix.h"
 #include "guilib/DirtyRegion.h"
-#include <stdint.h>
 #include <string>
 
 typedef enum _RenderingSystemType
@@ -37,12 +37,10 @@ typedef enum _RenderingSystemType
 } RenderingSystemType;
 
 /*
-*   CRenderSystemBase interface allows us to create the rendering engine we use.
-*   We currently have two engines: OpenGL and DirectX
-*   This interface is very basic since a lot of the actual details will go in to the derived classes
-*/
-
-typedef uint32_t color_t;
+ *   CRenderSystemBase interface allows us to create the rendering engine we use.
+ *   We currently have two engines: OpenGL and DirectX
+ *   This interface is very basic since a lot of the actual details will go in to the derived classes
+ */
 
 enum
 {
@@ -52,40 +50,6 @@ enum
   RENDER_CAPS_BGRA     = (1 << 3),
   RENDER_CAPS_BGRA_APPLE = (1 << 4)
 };
-
-enum
-{
-  RENDER_QUIRKS_MAJORMEMLEAK_OVERLAYRENDERER = 1 << 0,
-  RENDER_QUIRKS_YV12_PREFERED                = 1 << 1,
-  RENDER_QUIRKS_BROKEN_OCCLUSION_QUERY       = 1 << 2,
-};
-
-enum RENDER_STEREO_VIEW
-{
-  RENDER_STEREO_VIEW_OFF,
-  RENDER_STEREO_VIEW_LEFT,
-  RENDER_STEREO_VIEW_RIGHT,
-};
-
-enum RENDER_STEREO_MODE
-{
-  RENDER_STEREO_MODE_OFF,
-  RENDER_STEREO_MODE_SPLIT_HORIZONTAL,
-  RENDER_STEREO_MODE_SPLIT_VERTICAL,
-  RENDER_STEREO_MODE_ANAGLYPH_RED_CYAN,
-  RENDER_STEREO_MODE_ANAGLYPH_GREEN_MAGENTA,
-  RENDER_STEREO_MODE_ANAGLYPH_YELLOW_BLUE,
-  RENDER_STEREO_MODE_INTERLACED,
-  RENDER_STEREO_MODE_CHECKERBOARD,
-  RENDER_STEREO_MODE_HARDWAREBASED,
-  RENDER_STEREO_MODE_MONO,
-  RENDER_STEREO_MODE_COUNT,
-
-  // pseudo modes
-  RENDER_STEREO_MODE_AUTO = 100,
-  RENDER_STEREO_MODE_UNDEFINED = 999,
-};
-
 
 class CRenderSystemBase
 {
